+++ example_project/rapid_router_test_settings.py
@@ -171,7 +171,7 @@ try:
 except ImportError:
     pass
 
-from common.csp_config import *
+from codeforlife.legacy.csp_config import *
 
 AUTH_USER_MODEL = "user.User"