+++ game/end_to_end_tests/base_game_test.py
@@ -3,11 +3,15 @@ import socket
 import time
 from builtins import str
 
-from common.models import UserProfile
-from common.tests.utils.classes import create_class_directly
-from common.tests.utils.organisation import create_organisation_directly
-from common.tests.utils.student import create_school_student_directly
-from common.tests.utils.teacher import signup_teacher_directly
+from codeforlife.legacy.models import UserProfile
+from codeforlife.legacy.tests.utils.classes import create_class_directly
+from codeforlife.legacy.tests.utils.organisation import (
+    create_organisation_directly,
+)
+from codeforlife.legacy.tests.utils.student import (
+    create_school_student_directly,
+)
+from codeforlife.legacy.tests.utils.teacher import signup_teacher_directly
 from django.urls import reverse
 from game.models import Workspace