+++ game/templatetags/game/utils.py
@@ -1,6 +1,6 @@
 from django import template
 from django.template.defaultfilters import stringfilter
-from common.utils import using_two_factor
+from codeforlife.legacy.utils import using_two_factor
 
 register = template.Library()