+++ game/tests/test_level_editor.py
@@ -1,11 +1,15 @@
 import json
 from unittest.mock import patch
 
-from common.models import Teacher
-from common.tests.utils.classes import create_class_directly
-from common.tests.utils.organisation import create_organisation_directly
-from common.tests.utils.student import create_school_student_directly
-from common.tests.utils.teacher import signup_teacher_directly
+from codeforlife.legacy.models import Teacher
+from codeforlife.legacy.tests.utils.classes import create_class_directly
+from codeforlife.legacy.tests.utils.organisation import (
+    create_organisation_directly,
+)
+from codeforlife.legacy.tests.utils.student import (
+    create_school_student_directly,
+)
+from codeforlife.legacy.tests.utils.teacher import signup_teacher_directly
 from deploy import captcha
 from django.contrib.auth import get_user_model
 from django.core import mail