+++ game/tests/test_level_moderation.py
@@ -1,10 +1,12 @@
-from common.tests.utils.classes import create_class_directly
-from common.tests.utils.organisation import (
+from codeforlife.legacy.tests.utils.classes import create_class_directly
+from codeforlife.legacy.tests.utils.organisation import (
     create_organisation_directly,
     join_teacher_to_organisation,
 )
-from common.tests.utils.student import create_school_student_directly
-from common.tests.utils.teacher import signup_teacher_directly
+from codeforlife.legacy.tests.utils.student import (
+    create_school_student_directly,
+)
+from codeforlife.legacy.tests.utils.teacher import signup_teacher_directly
 from deploy import captcha
 from django.test.client import Client
 from django.test.testcases import TestCase